refactor: active block logic
