+++ web/components/gantt-chart/chart/views/bi-week.tsx
@@ -1,8 +1,9 @@
 import { FC } from "react";
+import { observer } from "mobx-react";
 // hooks
 import { useGanttChart } from "components/gantt-chart/hooks/use-gantt-chart";
 
-export const BiWeekChartView: FC<any> = () => {
+export const BiWeekChartView: FC<any> = observer(() => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { currentView, currentViewData, renderView } = useGanttChart();
 
@@ -50,4 +51,4 @@ export const BiWeekChartView: FC<any> = () => {
       </div>
     </>
   );
-};
+});
