+++ web/components/gantt-chart/chart/views/day.tsx
@@ -1,8 +1,9 @@
 import { FC } from "react";
+import { observer } from "mobx-react";
 // hooks
 import { useGanttChart } from "components/gantt-chart/hooks/use-gantt-chart";
 
-export const DayChartView: FC<any> = () => {
+export const DayChartView: FC<any> = observer(() => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { currentView, currentViewData, renderView } = useGanttChart();
 
@@ -50,4 +51,4 @@ export const DayChartView: FC<any> = () => {
       </div>
     </>
   );
-};
+});
