+++ web/components/gantt-chart/chart/views/quarter.tsx
@@ -1,8 +1,9 @@
 import { FC } from "react";
+import { observer } from "mobx-react";
 // hooks
 import { useGanttChart } from "components/gantt-chart/hooks/use-gantt-chart";
 
-export const QuarterChartView: FC<any> = () => {
+export const QuarterChartView: FC<any> = observer(() => {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const { currentView, currentViewData, renderView } = useGanttChart();
 
@@ -46,4 +47,4 @@ export const QuarterChartView: FC<any> = () => {
       </div>
     </>
   );
-};
+});
