+++ web/components/gantt-chart/sidebar/cycles/index.ts
@@ -1,0 +1,1 @@
+export * from "./sidebar";
