+++ web/components/gantt-chart/sidebar/issues/index.ts
@@ -1,0 +1,1 @@
+export * from "./sidebar";
