+++ web/components/gantt-chart/types/index.ts
@@ -22,33 +22,6 @@ export interface IBlockUpdateData {
 
 export type TGanttViews = "hours" | "day" | "week" | "bi_week" | "month" | "quarter" | "year";
 
-export interface ChartContextData {
-  currentView: TGanttViews;
-  currentViewData: ChartDataType | undefined;
-  renderView: any;
-  activeBlock: IGanttBlock | null;
-}
-
-export type ChartContextActionPayload =
-  | {
-      type: "CURRENT_VIEW";
-      payload: TGanttViews;
-    }
-  | {
-      type: "CURRENT_VIEW_DATA" | "RENDER_VIEW";
-      payload: ChartDataType | undefined;
-    }
-  | {
-      type: "PARTIAL_UPDATE";
-      payload: Partial<ChartContextData>;
-    };
-
-export interface ChartContextReducer extends ChartContextData {
-  scrollLeft: number;
-  updateScrollLeft: (scrollLeft: number) => void;
-  dispatch: (action: ChartContextActionPayload) => void;
-}
-
 // chart render types
 export interface WeekMonthDataType {
   key: number;
